refactor: start isolating FormSections into seperate component for resolving nested rerenders and validation focus issue
